+++ lib/repositories/limit-orders-repository.ts
@@ -6,7 +6,7 @@ import { DYNAMODB_TYPES } from '../config/dynamodb'
 import { OrderEntity } from '../entities'
 import { BaseOrdersRepository, MODEL_NAME } from './base'
 import { GenericOrdersRepository } from './generic-orders-repository'
-import { DutchIndexMapper } from './IndexMapper'
+import { DutchIndexMapper } from './IndexMappers/IndexMapper'
 import { getTableIndices, TABLE_NAMES } from './util'
 
 export class LimitOrdersRepository extends GenericOrdersRepository<string, string, null, OrderEntity> {
